Add fork command.
libvwm+:
Like '\t' tab_completion (that returns immediately), a ':' is like '\t'
but returns instead to the command line for further processing.
The fork command is such a case, that it also accepts a --command="com"command line argument.
When the command receives arguments the whole command should be double quoted, which
then interpreted as the value of the argument.

libvwm:
Add a vwm_frame_set_command (vwm_frame *, char *) that sets the frame
argvlist, by parsing the char * argument.
Instead the vwm_frame_set_argv() uses a char ** argument.