Remove the need for static variables to the extended library and to test applications.
libvwm+: hide the structure properties, and expose an initial api for
accessing its functionality and its properties.

As this is now in a state that the abstraction layer has been almost stabilized,
and there is enough functionality to be usable, update it to the second revision.
Actually this should be the first, as it is the first stable and usable product, after
the prototyping phase and the point zero, but nevermind.