Add a libvtach library that extend libvwm with attaching detaching capabilities. The algorithms for this library derived from the dtach utility.
dtach is a simple program that emulates the detach feature of screen.
dtach is (C)Copyright 2004-2016 Ned T. Crigler, and is under the GNU General
Public License.

This license is included in this directory.
Many thanks.

Included also a vtach utility that uses the library.

This library adds the detaching key binding, to the existing ones, which by
default is MODKEY-CTRL(d).

It can be later re-attached by using either "-a" or "--attach" from the
command line.

The "-s" or "--sockname=" is required on invocation.