libv: add a --send switch/option. This sends data to the specified socket through stdin and exits, e.g.,
create the socket:
  v --as=vim vim
detach with C-\ C-d

send the data:
  printf "aasdf\r" | v --as=vim --send
this will enter insert mode, and it will write asdfg and interpret a
carriage return by goind down one line

reattach:
  v --as=vim -a

Add libved as submodule and its corresponded targets in src/Makefile.

Replace "private" with "static" to libvwm, so it will allow building
the static targets with no naming conflicts. As a lesson "private" is not
as "static", so i've to reconsider using it from now on.

Use argparser from libved to parse the argument list.

libvwmed: add a :set command that for now gets a --log-file=[1|0] argument
that sets|unsets a log file for current frame respectively.