Continue to adjust the library for different scenarios.
However, this flexibility might also complicates the application interface at the
same time.

This is handled (normally) through documentation, or|and through code comments.
Both are missing. It is really hard to write documentation, not because is difficult
or because the desire is missing. No. It is hard because you get a feeling, that you
are stealing time from coding.