Restore and save an image of the current layout.
This is a preliminary addition, but that works well enough to be
usable, though there are unhandled conditions.

The initial setting is done through the set command:

  :set --image-name=name --save-image=1

Though there are other choises, this is the supported development intention,
as it makes more practical sense, since first is controlled through an inner
mechanism, and then it allows restoring the saved image, in a permanent way
at the shell invovation level, easily with:

  v --loadfile=name

This should restore the saved image.

But and for testing mostly purposes, it can work also that way:
  :set --image-file=/some/file --save-image=1
  :@save_image

or simply
  :@save_image --as=/some/file

But the image is saved automatically at the application deinitialization,
and if the --save-image=1 argument has been set before.
Using:

  :set --save-image=0

unsets the option.